pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "authlib>=1.6.5",
    "bson>=0.5.10",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "ruff>=0.12.12",
    "scalar-fastapi>=1.3.0",
    "sqlmodel>=0.0.24",
    "starlette>=0.47.3",
    "uvicorn>=0.35.0",
]


[tool.ruff]
line-length = 120
lint.select = ["E", "F", "W", "I"]
lint.ignore = ["E722", "E741", "C901", "E501"]
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
